thirdai_platform/requirements.txt
fastapi==0.109.2
gunicorn==20.1.0
jinja2==3.1.2
pdftitle==0.11 # 0.12 has has a conflict with pdfplumber from thirdai
pydantic>=2.1.0
PyMuPDF
requests<2.32.0
thirdai[neural_db_v2]>=0.9.17 --index-url https://download.pytorch.org/whl/cpu
aiohttp
alembic
apscheduler
Authlib
azure-identity
azure-mgmt-containerservice
azure-storage-blob
bcrypt
boto3
botocore
bs4
cohere
colorlog
ed25519
faker
fastapi[all]
fastapi_utils
fsspec
google-auth
google-auth-oauthlib
google-cloud-build
google-cloud-storage
hvac
httpx
langchain
langchain_openai
mmh3
msal
nltk
jinja2
pyyaml
python-keycloak
numpy
openai
pandas
prometheus-client
psycopg2-binary
PyJWT>=2.6.0
PyTrie
python-dotenv
python-multipart
pyyaml
sendgrid
sqlalchemy
starlette
tqdm
typing-extensions
unidecode
unstructured[pptx]
url-normalize
uvicorn[standard]
websockets
